Coder_Python/Projeto-python-MySQL/requirements.txt
flake8==4.0.1
mccabe==0.6.1
mysql-connector==2.2.9
pkg_resources==0.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
